tree.validNewItem: now accepts only one root
If any item already exists in the tree, no new roots will be
accepted. This means that items can connect to deleted items.

This solves the problem of having different DAGs after deleting an
item and then adding a new root. This causes trouble when trying
to merge from other sources that lack older versions.