merge_tree: support existing version numbers
If the version of a new local item (that is not in staging) is the
same as it's found parent, overwrite with a new version number.

This is to support drivers (like the mongodb adapter) that can not
easily determine if a new version is a confirmation of an async
update or a genuine new version and thus cannot drop h.v reliably.