browser: no conflict if new version equals store
In case a new merge item does not equal the expected previous version
but does equal the new version, don't write a conflict, just continue.

This can happen because writing merges is not atomic. I.e. in case the
use reloads while a new merge is written to the store, but not yet to
the DAG.